crates/sel4-initialize-tls: Refactor module structure
